Add forward_begin,end() methods on const_iterator
The SmartQuadtree<T> structure is now completely callback-free.

In order to iterate on pair, there is an iterating procedure through
forward_begin() and forward_end() for const_iterators. It is not yet
possible to non-const iterate on pairs of elements in the quadtree.

Note that forward_begin() builds a non symmetric relationship: if `a`
and `b` are candidates for being neighbours, either `a` is in
`b.forward_begin()` or `b` is in `a.forward_begin()` but not both.

Refs #2 #3 #4